multiple table crud project done

+++ 02.SQLite-Multiple-Tables-CRUD/app/src/main/java/com/hellohasan/sqlite_project/Features/SubjectCRUD/CreateSubject/Subject.java
@@ -6,7 +6,7 @@ public class Subject {
     private int code;
     private double credit;
 
-    public Subject(int id, String name, int code, double credit) {
+    public Subject(long id, String name, int code, double credit) {
         this.id = id;
         this.name = name;
         this.code = code;
